+++ src/directives/Tooltip/Tooltip.spec.js
@@ -1,6 +1,6 @@
 import { mount } from '@vue/test-utils'
 import * as Module from './Tooltip.ts'
-const Tooltip = Module.TooltipDirective
+const TooltipDirective = Module.TooltipDirective
 
 const factoryMount = (position = '', text = '') => {
   return mount(
@@ -14,7 +14,7 @@ const factoryMount = (position = '', text = '') => {
       },
       global: {
         directives: {
-          Tooltip: Tooltip
+          Tooltip: TooltipDirective
         }
       }
     }
@@ -24,19 +24,25 @@ const factoryMount = (position = '', text = '') => {
 describe('mounted hook', () => {
   test('does not append component when mounted', () => {
     const wrapper = factoryMount()
-    expect(wrapper.find('#tooltip-container').exists()).toBe(false)
+    const tooltipContainer = document.querySelector('#tooltip-container')
+
+    expect(tooltipContainer).toBe(null)
   })
 
   test('appends the component on mouseenter', async () => {
     const wrapper = factoryMount()
     await wrapper.trigger('mouseenter')
-    expect(wrapper.find('#tooltip-container').exists()).toBe(true)
+    const tooltipContainer = document.querySelector('#tooltip-container')
+
+    expect(tooltipContainer).not.toBe(null)
+    expect(tooltipContainer.classList.contains('tooltip')).toBe(true)
   })
 
   test('removes the component on mouseleave', async () => {
     const wrapper = factoryMount()
     await wrapper.trigger('mouseleave')
-    expect(wrapper.find('#tooltip-container').exists()).toBe(false)
+    const tooltipContainer = document.querySelector('#tooltip-container')
+    expect(tooltipContainer).toBe(null)
   })
 })
 
@@ -44,12 +50,15 @@ describe('arguments', () => {
   test('content is passed if provided', async () => {
     const wrapper = factoryMount('right', 'Hello')
     await wrapper.trigger('mouseenter')
-    expect(wrapper.find('.tooltip-content').text()).toBe('Hello')
+
+    const tooltipContainer = document.querySelector('#tooltip-container')
+    expect(tooltipContainer.textContent).toBe('Hello')
   })
 
   test('position is passed if provided ', async () => {
     const wrapper = factoryMount('right')
     await wrapper.trigger('mouseenter')
-    expect(wrapper.find('#tooltip-container').classes()).toContain('right')
+    const tooltipContainer = document.querySelector('#tooltip-container')
+    expect(tooltipContainer.classList.contains('right')).toBe(true)
   })
 })