+++ CHANGELOG.md
@@ -1,4 +1,6 @@
 - Update popup close button and improve typing on popup - [#78](https://github.com/PrefectHQ/miter-design/pull/78)
+- Fix issue with input background color - [#75](https://github.com/PrefectHQ/miter-design/pull/75)
+- Add popover component - [#67](https://github.com/PrefectHQ/miter-design/pull/67)
 - Update button styles - [#61](https://github.com/PrefectHQ/miter-design/pull/61)
 - Add grey-30 variable, fix placeholder variable - [#73](https://github.com/PrefectHQ/miter-design/pull/73)
 - Add placeholder prop to Input - [#70](https://github.com/PrefectHQ/miter-design/pull/70)